@@ -18,13 +18,15 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  flutter_svg: ^0.21.0-nullsafety.0
+  cupertino_icons: 
+  # flutter_svg: ^0.21.0-nullsafety.0
+  flutter_svg: ^2.0.9
   http: ^0.13.0
   infinite_scroll_pagination: ^3.0.1
   provider: ^5.0.0
