@@ -5,9 +5,9 @@ import '../size_config.dart';
 
 class DefaultButton extends StatelessWidget {
   const DefaultButton({
-    Key key,
-    this.text,
-    this.press,
+    Key? key,
+    required this.text,
+    required this.press,
   }) : super(key: key);
   final String text;
   final Function press;
@@ -17,10 +17,10 @@ class DefaultButton extends StatelessWidget {
     return SizedBox(
       width: double.infinity,
       height: getProportionateScreenHeight(56),
-      child: FlatButton(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-        color: kPrimaryColor,
-        onPressed: press,
+      child: TextButton(
+        // shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
+        // color: kPrimaryColor,
+        onPressed: () => press(),
         child: Text(
           text,
           style: TextStyle(
