@@ -5,8 +5,8 @@ import '../size_config.dart';
 
 class FormError extends StatelessWidget {
   const FormError({
-    Key key,
-    @required this.errors,
+    Key? key,
+    required this.errors,
   }) : super(key: key);
 
   final List<String> errors;
@@ -19,7 +19,7 @@ class FormError extends StatelessWidget {
     );
   }
 
-  Row formErrorText({String error}) {
+  Row formErrorText({required String error}) {
     return Row(
       children: [
         SvgPicture.asset(
