@@ -4,9 +4,9 @@ import '../size_config.dart';
 
 class RoundedIconBtn extends StatelessWidget {
   const RoundedIconBtn({
-    Key key,
-    @required this.icon,
-    @required this.press,
+    Key? key,
+    required this.icon,
+    required this.press,
     this.showShadow = false,
   }) : super(key: key);
 
@@ -30,7 +30,7 @@ class RoundedIconBtn extends StatelessWidget {
             ),
         ],
       ),
-      child: FlatButton(
+      child: MaterialButton(
         padding: EdgeInsets.zero,
         color: Colors.white,
         shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(50)),
