@@ -6,7 +6,7 @@ class Cart {
   final Product product;
   final int numOfItem;
 
-  Cart({@required this.product, @required this.numOfItem});
+  Cart({required this.product, required this.numOfItem});
 }
 
 // Demo data for our cart
