@@ -1,10 +1,15 @@
 import 'package:flutter/cupertino.dart';
 
-class DataAttr{
+class DataAttr {
   final int id;
 
-  DataAttr({@required this.id});
+  DataAttr({required this.id});
 }
 
 //demo category id;
-List<DataAttr> data=[DataAttr(id:1),DataAttr(id:2),DataAttr(id:3),DataAttr(id:4)];
+List<DataAttr> data = [
+  DataAttr(id: 1),
+  DataAttr(id: 2),
+  DataAttr(id: 3),
+  DataAttr(id: 4)
+];
