@@ -7,8 +7,8 @@ import '../../../size_config.dart';
 
 class CartCard extends StatelessWidget {
   const CartCard({
-    Key key,
-    @required this.cart,
+    Key? key,
+    required this.cart,
   }) : super(key: key);
 
   final Cart cart;
