@@ -5,9 +5,9 @@ import '../../../size_config.dart';
 
 class SectionTitle extends StatelessWidget {
   const SectionTitle({
-    Key key,
-    @required this.title,
-    @required this.press,
+    Key? key,
+    required this.title,
+    required this.press,
   }) : super(key: key);
 
   final String title;
@@ -16,7 +16,7 @@ class SectionTitle extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: const EdgeInsets.only(top:8.0),
+      padding: const EdgeInsets.only(top: 8.0),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.spaceBetween,
         children: [
