@@ -8,7 +8,7 @@ import 'section_title.dart';
 
 class SpecialOffers extends StatelessWidget {
   const SpecialOffers({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
@@ -32,17 +32,21 @@ class SpecialOffers extends StatelessWidget {
                 image: "assets/images/Image Banner 2.png",
                 category: "Smartphone",
                 numOfBrands: 18,
-                press: () {Navigator.pushNamed(
-            context,
-            ShowMore.routeName,
-            arguments: CategoryDetailsArgs(category: DataAttr(id:1)),
-          );},
+                press: () {
+                  Navigator.pushNamed(
+                    context,
+                    ShowMore.routeName,
+                    arguments: CategoryDetailsArgs(category: DataAttr(id: 1)),
+                  );
+                },
               ),
               SpecialOfferCard(
                 image: "assets/images/Image Banner 3.png",
                 category: "Fashion",
                 numOfBrands: 24,
-                press: () {Navigator.pushNamed(context, ShowMore.routeName);},
+                press: () {
+                  Navigator.pushNamed(context, ShowMore.routeName);
+                },
               ),
               SizedBox(width: getProportionateScreenWidth(20)),
             ],
@@ -55,11 +59,11 @@ class SpecialOffers extends StatelessWidget {
 
 class SpecialOfferCard extends StatelessWidget {
   const SpecialOfferCard({
-    Key key,
-    @required this.category,
-    @required this.image,
-    @required this.numOfBrands,
-    @required this.press,
+    Key? key,
+    required this.category,
+    required this.image,
+    required this.numOfBrands,
+    required this.press,
   }) : super(key: key);
 
   final String category, image;
