@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'constants.dart';
 
@@ -42,10 +43,13 @@ AppBarTheme appBarTheme() {
   return AppBarTheme(
     color: Colors.white,
     elevation: 0,
-    brightness: Brightness.light,
     iconTheme: IconThemeData(color: Colors.black),
-    textTheme: TextTheme(
-      headline6: TextStyle(color: Color(0XFF8B8B8B), fontSize: 18),
-    ),
+    systemOverlayStyle: SystemUiOverlayStyle.dark,
+    toolbarTextStyle: TextTheme(
+      titleLarge: TextStyle(color: Color(0XFF8B8B8B), fontSize: 18),
+    ).bodyMedium,
+    titleTextStyle: TextTheme(
+      titleLarge: TextStyle(color: Color(0XFF8B8B8B), fontSize: 18),
+    ).titleLarge,
   );
 }
