@@ -1,6 +1,8 @@
-package com.grb.magentomobile
+package com.example.shop_app
+
 
 import io.flutter.embedding.android.FlutterActivity
 
 class MainActivity: FlutterActivity() {
+
 }
