@@ -5,54 +5,55 @@
 import 'dart:convert';
 
 class CartRequestModals {
-    CartRequestModals({
-        this.cartItem,
-    });
+  CartRequestModals({
+    required this.cartItem,
+  });
 
-    final CartItem cartItem;
+  final CartItem cartItem;
 
-    factory CartRequestModals.fromRawJson(String str) => CartRequestModals.fromJson(json.decode(str));
+  factory CartRequestModals.fromRawJson(String str) =>
+      CartRequestModals.fromJson(json.decode(str));
 
-    String toRawJson() => json.encode(toJson());
+  String toRawJson() => json.encode(toJson());
 
-    factory CartRequestModals.fromJson(Map<String, dynamic> json) => CartRequestModals(
+  factory CartRequestModals.fromJson(Map<String, dynamic> json) =>
+      CartRequestModals(
         cartItem: CartItem.fromJson(json["cartItem"]),
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "cartItem": cartItem.toJson(),
-    };
+      };
 }
 
 class CartItem {
-    CartItem({
-        this.sku,
-        this.qty,
-        this.quoteId,
-    });
+  CartItem({
+    required this.sku,
+    required this.qty,
+    required this.quoteId,
+  });
 
-    final String sku;
-    final int qty;
-    final String quoteId;
+  final String sku;
+  final int qty;
+  final String quoteId;
 
-    factory CartItem.fromRawJson(String str) => CartItem.fromJson(json.decode(str));
+  factory CartItem.fromRawJson(String str) =>
+      CartItem.fromJson(json.decode(str));
 
-    String toRawJson() => json.encode(toJson());
+  String toRawJson() => json.encode(toJson());
 
-    factory CartItem.fromJson(Map<String, dynamic> json) => CartItem(
+  factory CartItem.fromJson(Map<String, dynamic> json) => CartItem(
         sku: json["sku"],
         qty: json["qty"],
         quoteId: json["quote_id"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "sku": sku,
         "qty": qty,
         "quote_id": quoteId,
-    };
+      };
 }
-
-
 
 //cart response modal
 
@@ -60,31 +61,32 @@ class CartItem {
 //
 //     final cartResponseModal = cartResponseModalFromJson(jsonString);
 
-
 class CartResponseModal {
-    CartResponseModal({
-        this.itemId,
-        this.sku,
-        this.qty,
-        this.name,
-        this.price,
-        this.productType,
-        this.quoteId,
-    });
+  CartResponseModal({
+    required this.itemId,
+    required this.sku,
+    required this.qty,
+    required this.name,
+    required this.price,
+    required this.productType,
+    required this.quoteId,
+  });
 
-    final int itemId;
-    final String sku;
-    final int qty;
-    final String name;
-    final double price;
-    final String productType;
-    final String quoteId;
+  final int itemId;
+  final String sku;
+  final int qty;
+  final String name;
+  final double price;
+  final String productType;
+  final String quoteId;
 
-    factory CartResponseModal.fromRawJson(String str) => CartResponseModal.fromJson(json.decode(str));
+  factory CartResponseModal.fromRawJson(String str) =>
+      CartResponseModal.fromJson(json.decode(str));
 
-    String toRawJson() => json.encode(toJson());
+  String toRawJson() => json.encode(toJson());
 
-    factory CartResponseModal.fromJson(Map<String, dynamic> json) => CartResponseModal(
+  factory CartResponseModal.fromJson(Map<String, dynamic> json) =>
+      CartResponseModal(
         itemId: json["item_id"],
         sku: json["sku"],
         qty: json["qty"],
@@ -92,9 +94,9 @@ class CartResponseModal {
         price: json["price"].toDouble(),
         productType: json["product_type"],
         quoteId: json["quote_id"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "item_id": itemId,
         "sku": sku,
         "qty": qty,
@@ -102,5 +104,5 @@ class CartResponseModal {
         "price": price,
         "product_type": productType,
         "quote_id": quoteId,
-    };
+      };
 }
