@@ -5,31 +5,33 @@
 import 'dart:convert';
 
 class CategoriesListModal {
-    CategoriesListModal({
-        this.id,
-        this.parentId,
-        this.name,
-        this.isActive,
-        this.position,
-        this.level,
-        this.productCount,
-        this.childrenData,
-    });
+  CategoriesListModal({
+    required this.id,
+    required this.parentId,
+    required this.name,
+    required this.isActive,
+    required this.position,
+    required this.level,
+    required this.productCount,
+    required this.childrenData,
+  });
 
-    int id;
-    int parentId;
-    String name;
-    bool isActive;
-    int position;
-    int level;
-    int productCount;
-    List<CategoriesListModal> childrenData;
+  int id;
+  int parentId;
+  String name;
+  bool isActive;
+  int position;
+  int level;
+  int productCount;
+  List<CategoriesListModal> childrenData;
 
-    factory CategoriesListModal.fromJson(String str) => CategoriesListModal.fromMap(json.decode(str));
+  factory CategoriesListModal.fromJson(String str) =>
+      CategoriesListModal.fromMap(json.decode(str));
 
-    String toJson() => json.encode(toMap());
+  String toJson() => json.encode(toMap());
 
-    factory CategoriesListModal.fromMap(Map<String, dynamic> json) => CategoriesListModal(
+  factory CategoriesListModal.fromMap(Map<String, dynamic> json) =>
+      CategoriesListModal(
         id: json["id"],
         parentId: json["parent_id"],
         name: json["name"],
@@ -37,10 +39,11 @@ class CategoriesListModal {
         position: json["position"],
         level: json["level"],
         productCount: json["product_count"],
-        childrenData: List<CategoriesListModal>.from(json["children_data"].map((x) => CategoriesListModal.fromMap(x))),
-    );
+        childrenData: List<CategoriesListModal>.from(
+            json["children_data"].map((x) => CategoriesListModal.fromMap(x))),
+      );
 
-    Map<String, dynamic> toMap() => {
+  Map<String, dynamic> toMap() => {
         "id": id,
         "parent_id": parentId,
         "name": name,
@@ -49,5 +52,5 @@ class CategoriesListModal {
         "level": level,
         "product_count": productCount,
         "children_data": List<dynamic>.from(childrenData.map((x) => x.toMap())),
-    };
+      };
 }
