@@ -4,94 +4,99 @@
 
 import 'dart:convert';
 
-ProductsModals productsModalsFromJson(String str) => ProductsModals.fromJson(json.decode(str));
+ProductsModals productsModalsFromJson(String str) =>
+    ProductsModals.fromJson(json.decode(str));
 
 String productsModalsToJson(ProductsModals data) => json.encode(data.toJson());
 
 class ProductsModals {
-    ProductsModals({
-        this.items,
-        this.searchCriteria,
-        this.totalCount,
-    });
+  ProductsModals({
+    required this.items,
+    required this.searchCriteria,
+    required this.totalCount,
+  });
 
-    List<Item> items;
-    SearchCriteria searchCriteria;
-    int totalCount;
+  List<Item> items;
+  SearchCriteria searchCriteria;
+  int totalCount;
 
-    factory ProductsModals.fromJson(Map<String, dynamic> json) => ProductsModals(
+  factory ProductsModals.fromJson(Map<String, dynamic> json) => ProductsModals(
         items: List<Item>.from(json["items"].map((x) => Item.fromJson(x))),
         searchCriteria: SearchCriteria.fromJson(json["search_criteria"]),
         totalCount: json["total_count"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "items": List<dynamic>.from(items.map((x) => x.toJson())),
         "search_criteria": searchCriteria.toJson(),
         "total_count": totalCount,
-    };
+      };
 }
 
 class Item {
-    Item({
-        this.id,
-        this.sku,
-        this.name,
-        this.attributeSetId,
-        this.price,
-        this.status,
-        this.visibility,
-        this.typeId,
-        this.createdAt,
-        this.updatedAt,
-        this.weight,
-        this.extensionAttributes,
-        this.productLinks,
-        this.options,
-        this.mediaGalleryEntries,
-        this.tierPrices,
-        this.customAttributes,
-    });
+  Item({
+    required this.id,
+    required this.sku,
+    required this.name,
+    required this.attributeSetId,
+    required this.price,
+    required this.status,
+    required this.visibility,
+    required this.typeId,
+    required this.createdAt,
+    required this.updatedAt,
+    required this.weight,
+    required this.extensionAttributes,
+    required this.productLinks,
+    required this.options,
+    required this.mediaGalleryEntries,
+    required this.tierPrices,
+    required this.customAttributes,
+  });
 
-    int id;
-    String sku;
-    String name;
-    int attributeSetId;
-    num price;
-    int status;
-    int visibility;
-    String typeId;
-    DateTime createdAt;
-    DateTime updatedAt;
-    int weight;
-    ExtensionAttributes extensionAttributes;
-    List<dynamic> productLinks;
-    List<dynamic> options;
-    List<MediaGalleryEntry> mediaGalleryEntries;
-    List<dynamic> tierPrices;
-    List<CustomAttribute> customAttributes;
+  int id;
+  String sku;
+  String name;
+  int attributeSetId;
+  num price;
+  int status;
+  int visibility;
+  String typeId;
+  DateTime createdAt;
+  DateTime updatedAt;
+  int weight;
+  ExtensionAttributes extensionAttributes;
+  List<dynamic> productLinks;
+  List<dynamic> options;
+  List<MediaGalleryEntry> mediaGalleryEntries;
+  List<dynamic> tierPrices;
+  List<CustomAttribute> customAttributes;
 
-    factory Item.fromJson(Map<String, dynamic> json) => Item(
+  factory Item.fromJson(Map<String, dynamic> json) => Item(
         id: json["id"],
         sku: json["sku"],
         name: json["name"],
         attributeSetId: json["attribute_set_id"],
-        price: json["price"]==null?0.0:json["price"],
+        price: json["price"] == null ? 0.0 : json["price"],
         status: json["status"],
         visibility: json["visibility"],
         typeId: json["type_id"],
         createdAt: DateTime.parse(json["created_at"]),
         updatedAt: DateTime.parse(json["updated_at"]),
         weight: json["weight"],
-        extensionAttributes: ExtensionAttributes.fromJson(json["extension_attributes"]),
+        extensionAttributes:
+            ExtensionAttributes.fromJson(json["extension_attributes"]),
         productLinks: List<dynamic>.from(json["product_links"].map((x) => x)),
         options: List<dynamic>.from(json["options"].map((x) => x)),
-        mediaGalleryEntries: List<MediaGalleryEntry>.from(json["media_gallery_entries"].map((x) => MediaGalleryEntry.fromJson(x))),
+        mediaGalleryEntries: List<MediaGalleryEntry>.from(
+            json["media_gallery_entries"]
+                .map((x) => MediaGalleryEntry.fromJson(x))),
         tierPrices: List<dynamic>.from(json["tier_prices"].map((x) => x)),
-        customAttributes: List<CustomAttribute>.from(json["custom_attributes"].map((x) => CustomAttribute.fromJson(x))),
-    );
+        customAttributes: List<CustomAttribute>.from(
+            json["custom_attributes"].map((x) => CustomAttribute.fromJson(x))),
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "sku": sku,
         "name": name,
@@ -106,92 +111,99 @@ class Item {
         "extension_attributes": extensionAttributes.toJson(),
         "product_links": List<dynamic>.from(productLinks.map((x) => x)),
         "options": List<dynamic>.from(options.map((x) => x)),
-        "media_gallery_entries": List<dynamic>.from(mediaGalleryEntries.map((x) => x.toJson())),
+        "media_gallery_entries":
+            List<dynamic>.from(mediaGalleryEntries.map((x) => x.toJson())),
         "tier_prices": List<dynamic>.from(tierPrices.map((x) => x)),
-        "custom_attributes": List<dynamic>.from(customAttributes.map((x) => x.toJson())),
-    };
+        "custom_attributes":
+            List<dynamic>.from(customAttributes.map((x) => x.toJson())),
+      };
 }
 
 class CustomAttribute {
-    CustomAttribute({
-        this.attributeCode,
-        this.value,
-    });
+  CustomAttribute({
+    required this.attributeCode,
+    this.value,
+  });
 
-    String attributeCode;
-    dynamic value;
+  String attributeCode;
+  dynamic value;
 
-    factory CustomAttribute.fromJson(Map<String, dynamic> json) => CustomAttribute(
+  factory CustomAttribute.fromJson(Map<String, dynamic> json) =>
+      CustomAttribute(
         attributeCode: json["attribute_code"],
         value: json["value"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "attribute_code": attributeCode,
         "value": value,
-    };
+      };
 }
 
 class ExtensionAttributes {
-    ExtensionAttributes({
-        this.websiteIds,
-        this.categoryLinks,
-    });
+  ExtensionAttributes({
+    required this.websiteIds,
+    required this.categoryLinks,
+  });
 
-    List<int> websiteIds;
-    List<CategoryLink> categoryLinks;
+  List<int> websiteIds;
+  List<CategoryLink> categoryLinks;
 
-    factory ExtensionAttributes.fromJson(Map<String, dynamic> json) => ExtensionAttributes(
+  factory ExtensionAttributes.fromJson(Map<String, dynamic> json) =>
+      ExtensionAttributes(
         websiteIds: List<int>.from(json["website_ids"].map((x) => x)),
-        categoryLinks: List<CategoryLink>.from(json["category_links"].map((x) => CategoryLink.fromJson(x))),
-    );
+        categoryLinks: List<CategoryLink>.from(
+            json["category_links"].map((x) => CategoryLink.fromJson(x))),
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "website_ids": List<dynamic>.from(websiteIds.map((x) => x)),
-        "category_links": List<dynamic>.from(categoryLinks.map((x) => x.toJson())),
-    };
+        "category_links":
+            List<dynamic>.from(categoryLinks.map((x) => x.toJson())),
+      };
 }
 
 class CategoryLink {
-    CategoryLink({
-        this.position,
-        this.categoryId,
-    });
+  CategoryLink({
+    required this.position,
+    required this.categoryId,
+  });
 
-    int position;
-    String categoryId;
+  int position;
+  String categoryId;
 
-    factory CategoryLink.fromJson(Map<String, dynamic> json) => CategoryLink(
+  factory CategoryLink.fromJson(Map<String, dynamic> json) => CategoryLink(
         position: json["position"],
         categoryId: json["category_id"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "position": position,
         "category_id": categoryId,
-    };
+      };
 }
 
 class MediaGalleryEntry {
-    MediaGalleryEntry({
-        this.id,
-        this.mediaType,
-        this.label,
-        this.position,
-        this.disabled,
-        this.types,
-        this.file,
-    });
+  MediaGalleryEntry({
+    required this.id,
+    required this.mediaType,
+    required this.label,
+    required this.position,
+    required this.disabled,
+    required this.types,
+    required this.file,
+  });
 
-    int id;
-    String mediaType;
-    String label;
-    int position;
-    bool disabled;
-    List<String> types;
-    String file;
+  int id;
+  String mediaType;
+  String label;
+  int position;
+  bool disabled;
+  List<String> types;
+  String file;
 
-    factory MediaGalleryEntry.fromJson(Map<String, dynamic> json) => MediaGalleryEntry(
+  factory MediaGalleryEntry.fromJson(Map<String, dynamic> json) =>
+      MediaGalleryEntry(
         id: json["id"],
         mediaType: json["media_type"],
         label: json["label"],
@@ -199,9 +211,9 @@ class MediaGalleryEntry {
         disabled: json["disabled"],
         types: List<String>.from(json["types"].map((x) => x)),
         file: json["file"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "media_type": mediaType,
         "label": label,
@@ -209,65 +221,68 @@ class MediaGalleryEntry {
         "disabled": disabled,
         "types": List<dynamic>.from(types.map((x) => x)),
         "file": file,
-    };
+      };
 }
 
 class SearchCriteria {
-    SearchCriteria({
-        this.filterGroups,
-        this.pageSize,
-    });
+  SearchCriteria({
+    required this.filterGroups,
+    required this.pageSize,
+  });
 
-    List<FilterGroup> filterGroups;
-    int pageSize;
+  List<FilterGroup> filterGroups;
+  int pageSize;
 
-    factory SearchCriteria.fromJson(Map<String, dynamic> json) => SearchCriteria(
-        filterGroups: List<FilterGroup>.from(json["filter_groups"].map((x) => FilterGroup.fromJson(x))),
+  factory SearchCriteria.fromJson(Map<String, dynamic> json) => SearchCriteria(
+        filterGroups: List<FilterGroup>.from(
+            json["filter_groups"].map((x) => FilterGroup.fromJson(x))),
         pageSize: json["page_size"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
-        "filter_groups": List<dynamic>.from(filterGroups.map((x) => x.toJson())),
+  Map<String, dynamic> toJson() => {
+        "filter_groups":
+            List<dynamic>.from(filterGroups.map((x) => x.toJson())),
         "page_size": pageSize,
-    };
+      };
 }
 
 class FilterGroup {
-    FilterGroup({
-        this.filters,
-    });
+  FilterGroup({
+    required this.filters,
+  });
 
-    List<Filter> filters;
+  List<Filter> filters;
 
-    factory FilterGroup.fromJson(Map<String, dynamic> json) => FilterGroup(
-        filters: List<Filter>.from(json["filters"].map((x) => Filter.fromJson(x))),
-    );
+  factory FilterGroup.fromJson(Map<String, dynamic> json) => FilterGroup(
+        filters:
+            List<Filter>.from(json["filters"].map((x) => Filter.fromJson(x))),
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "filters": List<dynamic>.from(filters.map((x) => x.toJson())),
-    };
+      };
 }
 
 class Filter {
-    Filter({
-        this.field,
-        this.value,
-        this.conditionType,
-    });
+  Filter({
+    required this.field,
+    required this.value,
+    required this.conditionType,
+  });
 
-    String field;
-    String value;
-    String conditionType;
+  String field;
+  String value;
+  String conditionType;
 
-    factory Filter.fromJson(Map<String, dynamic> json) => Filter(
+  factory Filter.fromJson(Map<String, dynamic> json) => Filter(
         field: json["field"],
         value: json["value"],
         conditionType: json["condition_type"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "field": field,
         "value": value,
         "condition_type": conditionType,
-    };
+      };
 }
