@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'components/body.dart';
 
 class CompleteProfileScreen extends StatelessWidget {
-  String email,pwd;
+  String? email, pwd;
   //CompleteProfileScreen(this.email,this.pwd);
   CompleteProfileScreen();
   static String routeName = "/complete_profile";
